+++ R/getMetricsFromYAML.R
@@ -22,19 +22,19 @@
 #' colnames(x)
 getMetricsFromYAML <- function(yaml) {
     assert(is.list(yaml))
-    cli_alert("Getting sample quality control metrics from YAML.")
+    alert("Getting sample quality control metrics from YAML.")
     data <- .sampleYAML(yaml, keys = c("summary", "metrics"))
     ## Early return on empty metrics (e.g. fast mode).
     if (!hasLength(data)) {
         ## nocov start
-        cli_alert_warning("No quality control metrics were calculated.")
+        alertWarning("No quality control metrics were calculated.")
         return(NULL)
         ## nocov end
     }
     ## Drop any metadata columns. Note we're also dropping the duplicate `name`
     ## column present in the metrics YAML.
     yamlFlatCols <- c("description", "genome_build", "sam_ref")
-    blacklist <- c(camelCase(yamlFlatCols), "name")
+    blacklist <- camelCase(c(yamlFlatCols, "name"), strict = TRUE)
     ## Drop blacklisted columns from the return.
     data <- data[, sort(setdiff(colnames(data), blacklist)), drop = FALSE]
     data