+++ R/package.R
@@ -3,4 +3,16 @@
 #' Base functions for bcbio R packages.
 #'
 #' @keywords internal
+#'
+#' @importMethodsFrom basejump coerce
+#'
+#' @importFrom basejump DataFrame alert alertInfo alertWarning as camelCase dl
+#'   import is localOrRemoteFile makeNames makeSampleData metadata<- na.omit new
+#'   packageName packageVersion printString realpath removeNA sanitizeNA tail ul
+#'   unlistToDataFrame
+#' @importFrom goalie allAreAtomic allAreFiles areDisjointSets
+#'   allAreMatchingRegex assert hasLength hasRownames isADirectory isAFile
+#'   isCharacter isMatchingRegex isInRange isInt isScalar isString isSubset
+#'   validNames
+#' @importFrom stringr str_match
 "_PACKAGE"