+++ NEWS.md
@@ -1,3 +1,12 @@
+## bcbioBase 0.6.17 (2021-02-10)
+
+### Minor changes
+
+- NAMESPACE updates, following basejump v0.14 release.
+- Hardened the YAML parser against nested metadata, and updated to use
+  `unlistToDataFrame` function defined in AcidPlyr, rather than deprecated
+  plyr `ldply` approach.
+
 ## bcbioBase 0.6.16 (2020-12-03)
 
 ### Minor changes